Divide viewController and view in mainViewController
