+++ Challange-1-iOS/Challange-1-iOS/Persistence/EmojiPersistence.swift
@@ -39,7 +39,7 @@ class EmojiPersistence {
             }
     }
 
-    func loadData() -> [Emoji] {
+    func fetchEmojisData() -> [Emoji] {
         var array: [NSManagedObject] = []
         var emojisArray: [Emoji] = []
