+++ Challange-1-iOS/Challange-1-iOS/Presentation/MainScreen/MainView.swift
@@ -1,0 +1,114 @@
+//
+//  MainView.swift
+//  Challange-1-iOS
+//
+//  Created by Daniel Silva on 11/4/22.
+//
+import UIKit
+//
+import RxCocoa
+import RxSwift
+
+class MainView: BaseGenericView {
+    private var verticalStackView: UIStackView
+    private var searchStackView: UIStackView
+    private var emojiContainer: UIView
+    private var btnRandomEmoji: UIButton
+    private var btnEmojisList: UIButton
+    private var btnAvatarsList: UIButton
+    private var btnAppleRepos: UIButton
+    private var btnSearch: UIButton
+    var searchBar: UISearchBar
+    var emojiImage: UIImageView
+    private var urlEmojiImage: String
+
+    var rxRandomEmojiTap: Observable<Void> { btnRandomEmoji.rx.tap.asObservable() }
+    var rxEmojiListTap: Observable<Void> { btnEmojisList.rx.tap.asObservable() }
+    var rxAvatarListTap: Observable<Void> { btnAvatarsList.rx.tap.asObservable() }
+    var rxAppleReposTap: Observable<Void> { btnAppleRepos.rx.tap.asObservable() }
+    var rxSearchTap: Observable<Void> { btnSearch.rx.tap.asObservable() }
+
+    required init() {
+        btnRandomEmoji = .init(type: .system)
+        btnEmojisList = .init(type: .system)
+        btnAvatarsList = .init(type: .system)
+        btnAppleRepos = .init(type: .system)
+        btnSearch = .init(type: .system)
+        searchBar = .init(frame: .zero)
+        emojiImage = .init(frame: .zero)
+        emojiContainer = .init(frame: .zero)
+        urlEmojiImage = .init()
+        searchStackView = .init(arrangedSubviews: [searchBar, btnSearch])
+        verticalStackView = .init(arrangedSubviews: [emojiContainer,
+                                                     btnRandomEmoji,
+                                                     btnEmojisList,
+                                                     searchStackView,
+                                                     btnAvatarsList,
+                                                     btnAppleRepos])
+        super.init()
+        createSubviews()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    func createSubviews() {
+        setUpViews()
+        addViewsToSuperview()
+        setUpConstraints()
+    }
+
+    // 1 - SetUp the views
+    private func setUpViews() {
+
+        backgroundColor = .appColor(name: .surface)
+        tintColor = .appColor(name: .secondary)
+
+        verticalStackView.axis = .vertical
+        searchStackView.axis = .horizontal
+        btnRandomEmoji.setTitle("Random List", for: .normal)
+        btnEmojisList.setTitle("Emojis List", for: .normal)
+        btnAvatarsList.setTitle("Avatars List", for: .normal)
+        btnAppleRepos.setTitle("Apple Repos", for: .normal)
+        btnSearch.setTitle("Search", for: .normal)
+
+        let buttonArray = [btnRandomEmoji, btnEmojisList, btnSearch, btnAvatarsList, btnAppleRepos]
+        buttonArray.forEach {
+            $0.configuration = .filled()
+        }
+    }
+
+    // 2 - Add to superview
+    private func addViewsToSuperview() {
+        emojiContainer.addSubview(emojiImage)
+
+        addSubview(verticalStackView)
+    }
+
+    // 3 - Set the constraints
+    private func setUpConstraints() {
+        btnSearch.translatesAutoresizingMaskIntoConstraints = false
+        verticalStackView.translatesAutoresizingMaskIntoConstraints = false
+        emojiImage.translatesAutoresizingMaskIntoConstraints = false
+        emojiContainer.translatesAutoresizingMaskIntoConstraints = false
+
+        NSLayoutConstraint.activate([
+            verticalStackView.centerYAnchor.constraint(equalTo: centerYAnchor),
+            verticalStackView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 30),
+            verticalStackView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -30),
+
+            emojiImage.topAnchor.constraint(equalTo: emojiContainer.topAnchor),
+            emojiImage.bottomAnchor.constraint(equalTo: emojiContainer.bottomAnchor, constant: -40),
+            emojiImage.leadingAnchor.constraint(equalTo: emojiContainer.leadingAnchor, constant: 100),
+            emojiImage.trailingAnchor.constraint(equalTo: emojiContainer.trailingAnchor, constant: -100),
+            emojiImage.widthAnchor.constraint(equalTo: emojiImage.heightAnchor, multiplier: 1)
+
+        ])
+
+        btnEmojisList.sizeToFit()
+
+        verticalStackView.spacing = 20
+        searchStackView.spacing = 20
+    }
+}
